Revert "Updated amount datatype"
This reverts commit 31302faa29bb4afcee38a2f9a88bc5e30819b7e0.

+++ municipal-services/fsm-calculator/src/main/java/org/egov/fsm/calculator/services/CalculationService.java
@@ -265,7 +265,7 @@ public class CalculationService {
 
 		// adding default amount as 0 when the pricing not mentioned manually
 		if (amount == null) {
-			amount = BigDecimal.valueOf(0);
+			amount = 0;
 			// throw new CustomException(CalculatorConstants.INVALID_PRICE,
 			// 		"Price not found in Billing Slab for the given propertyType and slumName");
 		}
